separate nonmainnet api to a new bin
