+++ e2e-tests/tests/api.rs
@@ -157,10 +157,11 @@ fn call_api() {
     assert!(rej.reject_message.contains("It's a trap!"));
 }
 
+// Test for the system API that are not enabled on the mainnet yet.
+// A canister Wasm module that imports the non-mainnet APIs would result in canister installation failure.
 #[test]
-fn call_env_var() {
-    let wasm = cargo_build_canister("api");
-    // env_var is not available on mainnet yet
+fn call_api_nonmainnet() {
+    let wasm = cargo_build_canister("api_nonmainnet");
     let pic = pic_base().with_nonmainnet_features(true).build();
     let canister_id = pic.create_canister();
     pic.add_cycles(canister_id, 100_000_000_000_000);
@@ -190,6 +191,7 @@ fn call_env_var() {
     )
     .unwrap();
 
+    // As of 2025-08, the env_var APIs are not available on mainnet yet.
     let res = pic
         .update_call(canister_id, sender, "call_env_var_count", vec![])
         .unwrap();
